test: move getModulePaths test to __tests__
